Split statement.go
